pyproject.toml
[project]
name = "langchain-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "langchain-google-vertexai>=3.2.0",
    "langchain-mistralai>=1.1.1",
    "langchain[google-genai,groq,mistralai]>=1.2.0",
    "python-dotenv>=1.2.1",
]
